+++ crates/swc/src/plugin.rs
@@ -42,9 +42,18 @@ struct RustPlugins {
 
 impl RustPlugins {
     #[cfg(feature = "plugin")]
-    fn apply(&mut self, mut n: Program) -> Result<Program, anyhow::Error> {
+    fn apply(&mut self, n: Program) -> Result<Program, anyhow::Error> {
         use anyhow::Context;
+        use swc_common::plugin::Serialized;
 
+        let mut serialized = Serialized::serialize(&n)?;
+
+        // Run plugin transformation against current program.
+        // We do not serialize / deserialize between each plugin execution but
+        // copies raw transformed bytes directly into plugin's memory space.
+        // Note: This doesn't mean plugin won't perform any se/deserialization: it
+        // still have to construct from raw bytes internally to perform actual
+        // transform.
         if let Some(plugins) = &self.plugins {
             for p in plugins {
                 let config_json = serde_json::to_string(&p.1)
@@ -52,17 +61,19 @@ impl RustPlugins {
 
                 let path = swc_plugin_runner::resolve::resolve(&p.0)?;
 
-                n = swc_plugin_runner::apply_js_plugin(
+                serialized = swc_plugin_runner::apply_js_plugin(
                     &p.0,
                     &path,
                     &mut self.plugin_cache,
                     &config_json,
-                    n,
+                    serialized,
                 )?;
             }
         }
 
-        Ok(n)
+        // Plugin transformation is done. Deserialize transformed bytes back
+        // into Program
+        Serialized::deserialize(&serialized)
     }
 }