+++ crates/swc_common/src/plugin.rs
@@ -8,38 +8,62 @@ use crate::{syntax_pos::Mark, SyntaxContext};
 use anyhow::Error;
 use std::any::type_name;
 
-/// Serialize ast to pass into plugins. This is being called from host side.
-/// TODO: https://github.com/swc-project/swc/issues/3167
-#[cfg(feature = "plugin-base")]
-pub fn serialize_for_plugin<T>(t: &T) -> Result<rkyv::AlignedVec, Error>
-where
-    T: rkyv::Serialize<rkyv::ser::serializers::AllocSerializer<512>>,
-{
-    rkyv::to_bytes::<_, 512>(t).map_err(|err| match err {
-        rkyv::ser::serializers::CompositeSerializerError::SerializerError(e) => e.into(),
-        rkyv::ser::serializers::CompositeSerializerError::ScratchSpaceError(e) => {
-            Error::msg("AllocScratchError")
-        }
-        rkyv::ser::serializers::CompositeSerializerError::SharedError(e) => {
-            Error::msg("SharedSerializeMapError")
-        }
-    })
+/// Wraps internal representation of serialized data. Consumers should not
+/// rely on specific details of byte format struct contains: it is
+/// strictly implementation detail which can change anytime.
+pub struct Serialized {
+    field: rkyv::AlignedVec,
 }
 
-/// TODO: https://github.com/swc-project/swc/issues/3167
-#[cfg(feature = "plugin-mode")]
-pub fn deserialize_for_plugin<T>(bytes: &[u8]) -> Result<T, Error>
-where
-    T: rkyv::Archive
-        + rkyv::with::DeserializeWith<<T as rkyv::Archive>::Archived, T, rkyv::Infallible>,
-{
-    use anyhow::Context;
-    use rkyv::Deserialize;
+#[cfg(feature = "plugin-base")]
+impl Serialized {
+    pub fn new_for_plugin(bytes: &[u8], len: i32) -> Serialized {
+        let mut vec = rkyv::AlignedVec::with_capacity(
+            len.try_into()
+                .expect("Cannot determine size of the serialized bytes"),
+        );
+        vec.extend_from_slice(&bytes);
+        Serialized { field: vec }
+    }
 
-    let archived = unsafe { rkyv::archived_root::<T>(&bytes[..]) };
+    pub fn from(vec: rkyv::AlignedVec) -> Serialized {
+        Serialized { field: vec }
+    }
 
-    archived
-        .deserialize(&mut rkyv::Infallible)
-        .map(|r: rkyv::with::With<T, T>| r.into_inner())
-        .with_context(|| format!("failed to deserialize `{}`", type_name::<T>()))
+    pub fn as_ref(&self) -> &rkyv::AlignedVec {
+        &self.field
+    }
+
+    pub fn serialize<W>(t: &W) -> Result<Serialized, Error>
+    where
+        W: rkyv::Serialize<rkyv::ser::serializers::AllocSerializer<512>>,
+    {
+        rkyv::to_bytes::<_, 512>(t)
+            .map(|v| Serialized::from(v))
+            .map_err(|err| match err {
+                rkyv::ser::serializers::CompositeSerializerError::SerializerError(e) => e.into(),
+                rkyv::ser::serializers::CompositeSerializerError::ScratchSpaceError(e) => {
+                    Error::msg("AllocScratchError")
+                }
+                rkyv::ser::serializers::CompositeSerializerError::SharedError(e) => {
+                    Error::msg("SharedSerializeMapError")
+                }
+            })
+    }
+
+    pub fn deserialize<W>(bytes: &Serialized) -> Result<W, Error>
+    where
+        W: rkyv::Archive,
+        W::Archived: rkyv::Deserialize<W, rkyv::Infallible>,
+    {
+        use anyhow::Context;
+        use rkyv::Deserialize;
+
+        let bytes = &bytes.field;
+        let archived = unsafe { rkyv::archived_root::<W>(&bytes[..]) };
+
+        archived
+            .deserialize(&mut rkyv::Infallible)
+            .with_context(|| format!("failed to deserialize `{}`", type_name::<W>()))
+    }
 }